refactor(engine): extract daemon runner types and schemas to src/daemon/runner/
Runner-internal types and tool schemas extracted from workflow-runner.ts:
- runner/runner-types.ts: PreAgentSession, PreAgentSessionResult, AgentReadySession,
  SessionOutcome, FinalizationContext, WORKTREES_DIR
- runner/tool-schemas.ts: getSchemas() with all JSON Schema definitions for daemon tools

Architecture test enforces runner/ does not import runWorkflow from workflow-runner.ts
at runtime (prevents circular dependency).

Note: constructTools(), buildPreAgentSession(), buildAgentReadySession(), runAgentLoop(),
buildAgentCallbacks(), and buildTurnEndSubscriber() remain in workflow-runner.ts pending
a follow-on refactor that introduces runWorkflow injection to eliminate the circular dep.
TurnEndSubscriberContext remains inline alongside its corresponding function.

378/378 test files, 5860 tests pass. Build clean.
